+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+- moves to bash-parser
+
 # 0.0.0
 
 - Can profile very basic shell script